Fix for unequip able items while using EngBags
Right clicking an unequip able item causes the item to get stuck in a
locked state. While reading through the EngInventory code the
ITEM_LOCK_CHANGED event is commented out. So technically this is an
EngBags bug. I added this code block to EquipColor instead which forces
EngBags to update all the inventory objects freeing the locked objects
state allowing EquipColor to recolor the item.